Upload setup photo and multi-view system.
Panels: (a) Picture of the experimental setup, where the tank filled with fluid between the two sides of the transparent plastic belt is visible (1), together with the motion transmission system (2) and the two cameras, lookingatthe(x,y)flow-gradientplane(3)andatthe(x,z),flow-vorticity plane (4). (b) Sketch of the output given by the dual-camera video- recording system. The reference frame is defined at the centre of the particle. The azimuth and polar angles, φ and θ, as well as the projected angle, λ, and the three components of the Axes-Aligned Bounding Box, B, are also represented. The two recorded frames are displayed on the corre- sponding (x,y) flow-gradient and (x,z) flow-vorticity planes, to appreci- ate the contrast between the particle projections and the background as well as the detected particle contours and Axes-Aligned Bounding Boxes.