Experimental setup
Panel (a): realistic rendering of the linear shear cell deployed in our experiments. The small tank is displayed (1), together with its lid (2), from which the two cylinders hang. The first is free to rotate (3), being coupled to a transmission shaft (4) through a rolling bearing. The second is fixed (5). Between them, a transparent plastic belt made of Mylar is kept under tension (6). Neither the laser-cut holes in the belt nor the gear glued on the rotating cylinder is visualized for sake of simplicity. The two cameras are also portrayed: one looking at the flow-gradient (x, y) plane (7), and the other focused on the flow-vorticity (x,z) plane (8). The scale is also reported in blue.
Panel (b): picture of the experimental setup, where the tank filled with fluid between the two sides of the transparent plastic belt is visible (1), together with the motion transmission system (9) and the two cameras, looking at the(x,y) flow-gradient plane (7) and at the (x,z),flow-vorticity plane (8). 
Panel (c): sketch of the output given by the dual-camera video-recording system. The reference frame is defined at the centre of the particle. The azimuth and polar angles, φ and θ, as well as the projected angle, λ, and the three components of the Axes-Aligned Bounding Box, B, are also represented. The two recorded frames are displayed on the corresponding (x,y) flow-gradient and (x,z) flow-vorticity planes, to appreciate the contrast between the particle projections and the background as well as the detected particle contours and Axes-Aligned Bounding Boxes.